Add zenoss.winrm.Software modeler plugin.
This modeler plugin is functionality equivalent to the
zenoss.wmi.SoftwareMap modeler from the WindowsMonitor ZenPack. The only
difference is that the software manufacturer is also modeler.

Rob, and the Internet in general suggest that to get an accurate list of
installed software you should also query registry keys under the
following root.

HKLM:\Software\Microsoft\Windows\CurrentVersion\Uninstall

In the future this modeler plugin could be extended to query the
Win32_Product and the registry and merge the results.