Use corresponding txwinrm branch if possible.
This changes the ZenPack build process. Previously the txwinrm master
branch would always be used in this ZenPack. That led to problems when
txwinrm was actively being updated to support changes in the ZenPack.
With this change the txwinrm master branch will only be used when the
ZenPack is built from its own master branch. Otherwise the build will
attempt to checkout the txwinrm branch that matches the ZenPack's
current branch and fallback to txwinrm's develop branch if that fails.