zWin(User|Password) -> zWinRM(User|Password).
This is necessary to support the use case where you want to monitor the
same device partially with the old WindowsMonitor DCOM/RPC support and
this ZenPack's WinRM support with a domain user. The old zWinUser format
has to be "DOMAIN\Username" and the new zWinRMUser format must be
"username@example.com".