Fix collection failure on empty results stdout.
It appears that there are occasions when the onReceive callback gets
called with no stdout data. The only way I've been able to reproduce
this is to forcibly kill the Get-Counter command on the Windows server.
When I did that it caused a traceback into the bowls of Twisted that
would prevent further perfmon collection on the device.