Collect perfmon data 1 cycle earlier.
When we ask for data, we know that we won't get the data back until now
+ cycletime. This introduces a race condition on the next collection
interval where the data won't yet be available for a fraction of a
second. This means we have to wait another collection interval before
the data will be available. Doing an asynchronous sleep for 1 second
here mitigates the race condition and greatly increases the chances that
we'll get data a full collection interval earlier.