Fix leaking operations in shell datasource.
The "Windows Shell" datasource was attempting to reuse the same shell to
run many commands. The biggest problem was that it wasn't done very well
and it was leaving a lot of "active operations" on the Windows server
that were eating up resources and eventually causing Zenoss to be denied
WS-Management access. The other problem is that commands couldn't be run
in parallel.

This change stops trying to share a shell between commands and
simplifies the datasource considerably. It also properly cleans up after
itself on the Windows server.