Add zWinRMServerName property.
Consolidate code for creating ConnectionInfo instances. This creates a
consistent approach to testing for and reporting on potential
configuration issues.

This change also adds support for calling back to zWinUser and
zWinPassword when zWinRMUser and zWinRMPassword haven't been set. This
was done to help users migrate from the previous Windows ZenPack and to
potentially smooth over issues where users are using past experience or
older troubleshooting guides available on the Internet to configure the
new Windows ZenPack.

Fixes ZEN-9712, ZEN-9398.