Fix partial subclassing of platform model classes
When zenpacklib is used with both zenpacklib.Component and a platform
component model classes such as IpInterface as base classes, some of the
platform model class' functionality is list. Specifically in this case
the IpInterface.index_object() method was never being called when
Windows Interface.index_object() was called. There's important
functionality in IpInterface.index_object() that must be used.

This fix enhances zenpacklib to provide a proxy base type for all
platform component model classes. This allows single-inheritence to be
used in zenpack.yaml, which greatly simplifies the ZenPack's job because
zenpacklib can take on the responsibility of making sure base class
functionality is used when appropriate.

Fixes ZEN-25559.