Use get-clusterresource so we can get the id/ownernode on 2008
Fixes ZPS-2015

the MSCluster_Resource WMI class does not contain the Id and OwnerNode
attributes on 2008.  They're only on 2012, 2016+.  get-clusterresource
will return a class that does contain the required attributes.  we'll
keep our wmi loops however as this seems to be a more accepted method
of finding cluster disks and their sizes.