double check corrupt counters
Fixes ZPS-2383

This will double check that counters are actually corrupt so that we don't have a false negative.  For example, get-counter could succeed in showing a sample for a counter, but there could be something written to stderr.  for now, let's just log the error as debug.  Added unit test