Connect to correct node to pull database, job, backups
Fixes ZPS-3461

be sure to connect to the correct device in the conn_info to pull info.  get ip address of node to do this.  Updated unit test for cluster.  query db to get version for clustered sql server.  we should already have version for standalone server so no need to query for it.