fix service status issue, kerb ticket expiry
Fixes ZPS-3216

txwinrm reviewed in https://github.com/zenoss/txwinrm/pull/153.  it handles edge case where a ticket expires in the middle of establishing a connection.  also we need to send a clear to /Status.

Additional change for ZPS-3461.  if no ip address returned, then we need to just use the ownernode hostname.