raise event for possible powershell config issue
Fixes ZPS-3517

if the PSModulePath is wrong or contains a UNC path, powershell modules could fail to load and we'd see errors about simple powershell cmdlets not working.  this will raise an event giving user info to be sure their path is right.  also documented this in a troubleshooting section