do not restart() on receiveFail
Fixes ZPS-3377

when receive fails on a device reboot, we were restarting the entire process.  this is unnecessary and the reason we see tons of error messages.  we will just wait until the next cycle to start collection over.  handling certain error messages and sending them to debug with an explanation that they are possibly due to a device reboot.

updated releases sections for proper formatting of releases