attempt to stop and reset on receive failures
Fixes ZPS-4473

if we are failing to receive, this can be due to many factors.  could be a simple OperationTimeout, which means counters have not reported, so we will retry.  on a timeout error we should retry to receive.  anything else we will try to stop the command and reset ourselves into a better state.