account for cluster shared volumes freespace and status
Fixes ZPS-4612

Cluster shared volumes are not returned from get-clusterresource so we can't get disk state that way.  we can map the state to our disk states however.  captured customer data in a pickle to create a unit test to ensure we are setting the correct state and updating freespace