both cluster group and sql server can failover separately
Fixes ZPS-4707

Since those services can failover separately we should keep tabs on both of them. there could be more in the future we should keep an eye on, but for now these two will suffice.  updated unit test to account for cluster output and to check for owner change events.  both could fail over and send separate events.  this shouldn't cause multiple remodels as we're checking for an existing remodel job.