take into account lack of get-disk/get-partition on win2008
fixes ZPS-5582

when modeling, we will skip getting the disk number for win 2008 devices.  determining a workaround with wmi objects is not a useful endeavor because there are no direct associations.  also, when monitoring, we only need to worry about getting the id, freespace, and state of the csv.  we should have also been ignoring the csvs when getting 'Physical Disk' resources.  updated doc to indicate that an no disk number will be reported for the csvs on win 2008